4A3E0

+++ src/373A0.h
@@ -75,7 +75,7 @@ extern u8* D_800FC6C0[3];
 extern unk_D_800FC6CC* D_800FC6CC;
 extern u8* D_800FC6D0;
 extern unk_D_800FC6D4* D_800FC6D4;
-extern u8* D_800FC6D8;
+extern s16* D_800FC6D8;
 extern u8* D_800FC6DC;
 extern u32* D_800FC6E0;
 extern File1Header* D_800FC6E4;
